+++ build/ecsy.module.js
@@ -7,9 +7,7 @@ class SystemManager {
   }
 
   registerSystem(System, attributes) {
-    if (
-      this._systems.find(s => s.constructor.name === System.name) !== undefined
-    ) {
+    if (this.getSystem(System) !== undefined) {
       console.warn(`System '${System.name}' already registered.`);
       return this;
     }
@@ -339,7 +337,7 @@ class Entity {
 
     this.alive = false;
 
-    //if there are state components on a entity, it can't be removed
+    //if there are state components on a entity, it can't be removed completely
     this.numStateComponents = 0;
   }
 
@@ -392,8 +390,8 @@ class Entity {
     return this;
   }
 
-  removeComponent(Component, forceRemove) {
-    this._world.entityRemoveComponent(this, Component, forceRemove);
+  removeComponent(Component, forceImmediate) {
+    this._world.entityRemoveComponent(this, Component, forceImmediate);
     return this;
   }
 
@@ -422,8 +420,8 @@ class Entity {
     return false;
   }
 
-  removeAllComponents(forceRemove) {
-    return this._world.entityRemoveAllComponents(this, forceRemove);
+  removeAllComponents(forceImmediate) {
+    return this._world.entityRemoveAllComponents(this, forceImmediate);
   }
 
   // EXTRAS
@@ -437,8 +435,8 @@ class Entity {
     this._components = {};
   }
 
-  remove(forceRemove) {
-    return this._world.removeEntity(this, forceRemove);
+  remove(forceImmediate) {
+    return this._world.removeEntity(this, forceImmediate);
   }
 }
 
@@ -677,7 +675,14 @@ class EntityManager {
    * @param {Object} values Optional values to replace the default attributes
    */
   entityAddComponent(entity, Component, values) {
-    if (~entity._ComponentTypes.indexOf(Component)) return;
+    if (~entity._ComponentTypes.indexOf(Component)) {
+      console.warn(
+        "Component type already exists on entity.",
+        entity,
+        Component
+      );
+      return;
+    }
 
     entity._ComponentTypes.push(Component);
 
@@ -1071,6 +1076,9 @@ var pjson = {
 
 const Version = pjson.version;
 
+const hasWindow = typeof window !== "undefined";
+const hasCustomEvent = typeof CustomEvent !== "undefined";
+
 class World {
   constructor() {
     this.componentsManager = new ComponentManager(this);
@@ -1081,7 +1089,7 @@ class World {
 
     this.eventQueues = {};
 
-    if (typeof CustomEvent !== "undefined") {
+    if (hasWindow && hasCustomEvent) {
       var event = new CustomEvent("ecsy-world-created", {
         detail: { world: this, version: Version }
       });
@@ -1341,7 +1349,11 @@ function Not(Component) {
   };
 }
 
-class Component {}
+class Component {
+  copy() {}
+  reset() {}
+  clear() {}
+}
 
 Component.isComponent = true;
 
@@ -1600,6 +1612,7 @@ function injectScript(src, onLoad) {
 }
 
 /* global Peer */
+const hasWindow$1 = typeof window !== "undefined";
 
 function hookConsoleAndErrors(connection) {
   var wrapFunctions = ["error", "warning", "log"];
@@ -1617,6 +1630,7 @@ function hookConsoleAndErrors(connection) {
     }
   });
 
+  if(!hasWindow$1) return;
   window.addEventListener("error", error => {
     connection.send({
       method: "error",
@@ -1654,6 +1668,7 @@ function includeRemoteIdHTML(remoteId) {
 }
 
 function enableRemoteDevtools(remoteId) {
+  if(!hasWindow$1) return;
   window.generateNewCode = () => {
     window.localStorage.clear();
     remoteId = generateId(6);
@@ -1739,11 +1754,13 @@ function enableRemoteDevtools(remoteId) {
   );
 }
 
-const urlParams = new URLSearchParams(window.location.search);
+if (hasWindow$1) {
+  const urlParams = new URLSearchParams(window.location.search);
 
-// @todo Provide a way to disable it if needed
-if (urlParams.has("enable-remote-devtools")) {
-  enableRemoteDevtools();
+  // @todo Provide a way to disable it if needed
+  if (urlParams.has("enable-remote-devtools")) {
+    enableRemoteDevtools();
+  }
 }
 
 export { Component, Not, System, SystemStateComponent, TagComponent, Types, Version, World, createComponentClass, createType, enableRemoteDevtools };
